5396: rename search-list-page to card-list-page
