+++ src/apps/search-result/search-result.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { useDeepCompareEffect } from "react-use";
 import SearchResultHeader from "../../components/search-bar/search-result-header/SearchResultHeader";
 import usePager from "../../components/result-pager/use-pager";
-import SearchResultList from "../../components/search-result-list/SearchResultList";
+import SearchResultList from "../../components/card-item-list/SearchResultList";
 import {
   FacetField,
   SearchWithPaginationQuery,
@@ -166,7 +166,7 @@ const SearchResult: React.FC<SearchResultProps> = ({ q, pageSize }) => {
   }, [clearFilter]);
 
   return (
-    <div className="search-result-page">
+    <div className="card-list-page">
       <SearchResultHeader hitcount={hitcount} q={q} />
       <FacetLine q={q} />
       {campaignData && campaignData.data && (
