+++ src/components/card-item-list/SearchResultList.tsx
@@ -1,8 +1,8 @@
 import React, { memo } from "react";
 import { dataIsNotEmpty, getCoverTint } from "../../core/utils/helpers/general";
 import { Work } from "../../core/utils/types/entities";
-import SearchResultListItem from "./search-result-list-item/search-result-list-item";
-import SearchResultListItemSkeleton from "./search-result-list-item/search-result-list-item-skeleton";
+import SearchResultListItem from "./search-result-list-item/card-list-item";
+import SearchResultListItemSkeleton from "./search-result-list-item/card-list-item-skeleton";
 
 export interface SearchResultListProps {
   resultItems: Work[];
@@ -12,7 +12,7 @@ const SearchResultList: React.FC<SearchResultListProps> = ({ resultItems }) => {
   const worksAreLoaded = dataIsNotEmpty(resultItems);
 
   return (
-    <ul className="search-result-page__list my-32" data-cy="search-result-list">
+    <ul className="card-list-page__list my-32" data-cy="search-result-list">
       {/*
           Show skeleton search result items if no data is available yet.
           We'll show 5 items which should cover most screens.
