+++ src/components/card-item-list/search-result-list-item/card-list-item-cover.tsx
